Create String Service
