+++ Want2LearnBase/Want2LearnBase.csproj
@@ -47,6 +47,7 @@
     <Compile Include="MatrixService.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StringService.cs" />
     <Compile Include="SwapService.cs" />
   </ItemGroup>
   <ItemGroup>
